Move some controller code into Admin\ namespace
