+++ src/Controller/PackageController.php
@@ -48,7 +48,6 @@ use App\Security\Voter\PackageActions;
 use App\SecurityAdvisory\GitHubSecurityAdvisoriesSource;
 use App\Service\GitHubUserMigrationWorker;
 use App\Service\Scheduler;
-use App\Service\Spam\SpamClassifier;
 use App\Util\Killswitch;
 use Composer\MetadataMinifier\MetadataMinifier;
 use Composer\Package\Version\VersionParser;
@@ -101,7 +100,6 @@ class PackageController extends Controller
         /** @var AwsMetadata */
         private array $awsMetadata,
         private EventDispatcherInterface $eventDispatcher,
-        private SpamClassifier $spamClassifier,
     ) {
     }
 
@@ -398,99 +396,6 @@ class PackageController extends Controller
             'meta' => $this->getPackagesMetadata($this->favoriteManager, $this->downloadManager, $providers),
             'paginate' => false,
         ]);
-    }
-
-    #[Route(path: '/admin/spam', name: 'view_spam', defaults: ['_format' => 'html'], methods: ['GET'])]
-    public function viewSpamAction(Request $req, CsrfTokenManagerInterface $csrfTokenManager): Response
-    {
-        if (!$this->getUser() || !$this->isGranted('ROLE_ANTISPAM')) {
-            throw new NotFoundHttpException();
-        }
-
-        $page = max(1, $req->query->getInt('page', 1));
-
-        $repo = $this->getEM()->getRepository(Package::class);
-        $count = $repo->getSuspectPackageCount();
-        $packages = $repo->getSuspectPackages(($page - 1) * 50, 50);
-
-        $paginator = new Pagerfanta(new FixedAdapter($count, $packages));
-        $paginator->setNormalizeOutOfRangePages(true);
-        $paginator->setMaxPerPage(50);
-        $paginator->setCurrentPage($page);
-
-        $data['packages'] = $paginator;
-        $data['count'] = $count;
-        $data['meta'] = $this->getPackagesMetadata($this->favoriteManager, $this->downloadManager, $data['packages']);
-        $data['meta']['spamScores'] = $this->computeSpamScores($repo, $packages);
-        $data['markSafeCsrfToken'] = $csrfTokenManager->getToken('mark_safe');
-
-        $vendorRepo = $this->getEM()->getRepository(Vendor::class);
-        $verified = [];
-        foreach ($packages as $pkg) {
-            $dls = $data['meta']['downloads'][$pkg['id']] ?? 0;
-            $vendor = Preg::replace('{/.*$}', '', $pkg['name']);
-            if ($dls > 10 && !\in_array($vendor, $verified, true)) {
-                $vendorRepo->verify($vendor);
-                $this->addFlash('success', 'Marked '.$vendor.' with '.$dls.' downloads.');
-                $verified[] = $vendor;
-            }
-        }
-
-        if ($verified) {
-            return $this->redirectToRoute('view_spam');
-        }
-
-        return $this->render('admin/spam.html.twig', $data);
-    }
-
-    /**
-     * Runs the spam classifier over the listed packages so moderators can see, per package, how
-     * spammy the model thinks it is and whether it would be auto-cleared. No-op (empty map) when no
-     * model is deployed. The score is cheap enough to compute inline for a page of results.
-     *
-     * @param array<array{id: int, name: string, description: string|null}> $packages
-     *
-     * @return array<int, array{metadata: float, readme: float|null, safe: bool}>
-     */
-    private function computeSpamScores(PackageRepository $repo, array $packages): array
-    {
-        if (\count($packages) === 0 || !$this->spamClassifier->isModelAvailable()) {
-            return [];
-        }
-
-        $ids = array_values(array_map(static fn (array $pkg) => $pkg['id'], $packages));
-        $tagsById = $repo->getTagsByPackageIds($ids);
-        $readmesById = $repo->getReadmeContentsByPackageIds($ids);
-
-        $scores = [];
-        foreach ($packages as $pkg) {
-            $scores[$pkg['id']] = $this->spamClassifier->evaluate(
-                $pkg['name'],
-                $pkg['description'],
-                $tagsById[$pkg['id']] ?? [],
-                $readmesById[$pkg['id']] ?? null,
-            );
-        }
-
-        return $scores;
-    }
-
-    #[IsGranted('ROLE_ANTISPAM')]
-    #[Route(path: '/admin/spam/nospam', name: 'mark_nospam', defaults: ['_format' => 'html'], methods: ['POST'])]
-    public function markSafeAction(Request $req): RedirectResponse
-    {
-        /** @var string[] $vendors */
-        $vendors = array_filter($req->request->all('vendor'), static fn ($vendor) => $vendor !== '' && $vendor !== null);
-        if (!$this->isCsrfTokenValid('mark_safe', $req->request->getString('token'))) {
-            throw new BadRequestHttpException('Invalid CSRF token');
-        }
-
-        $repo = $this->getEM()->getRepository(Vendor::class);
-        foreach ($vendors as $vendor) {
-            $repo->verify($vendor);
-        }
-
-        return $this->redirectToRoute('view_spam');
     }
 
     #[IsGranted('ROLE_ADMIN')]
